refactor(Q-3): modulariza pastas das entidades por dto e model
