+++ src/main/java/br/com/alura/AluraFake/course/CourseController.java
@@ -1,7 +1,11 @@
 package br.com.alura.AluraFake.course;
 
+import br.com.alura.AluraFake.course.dto.CourseListItemDTO;
+import br.com.alura.AluraFake.course.dto.NewCourseDTO;
+import br.com.alura.AluraFake.course.model.Course;
 import br.com.alura.AluraFake.task.TaskRepository;
 import br.com.alura.AluraFake.user.*;
+import br.com.alura.AluraFake.user.model.User;
 import br.com.alura.AluraFake.util.ErrorItemDTO;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
