+++ src/main/java/br/com/alura/AluraFake/course/dto/CourseListItemDTO.java
@@ -1,4 +1,7 @@
-package br.com.alura.AluraFake.course;
+package br.com.alura.AluraFake.course.dto;
+
+import br.com.alura.AluraFake.course.model.Course;
+import br.com.alura.AluraFake.course.model.CourseStatus;
 
 import java.io.Serializable;
