+++ src/main/java/br/com/alura/AluraFake/course/model/CourseStatus.java
@@ -1,4 +1,4 @@
-package br.com.alura.AluraFake.course;
+package br.com.alura.AluraFake.course.model;
 
 public enum CourseStatus {
     BUILDING,
