+++ src/main/java/br/com/alura/AluraFake/task/dto/NewSingleChoiceTaskDTO.java
@@ -1,6 +1,7 @@
-package br.com.alura.AluraFake.task;
+package br.com.alura.AluraFake.task.dto;
 
-import br.com.alura.AluraFake.alternative.NewAlternativeDTO;
+import br.com.alura.AluraFake.alternative.dto.NewAlternativeDTO;
+import br.com.alura.AluraFake.task.model.TaskType;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
