+++ src/main/java/br/com/alura/AluraFake/task/model/Task.java
@@ -1,7 +1,7 @@
-package br.com.alura.AluraFake.task;
+package br.com.alura.AluraFake.task.model;
 
-import br.com.alura.AluraFake.alternative.Alternative;
-import br.com.alura.AluraFake.course.Course;
+import br.com.alura.AluraFake.alternative.model.Alternative;
+import br.com.alura.AluraFake.course.model.Course;
 import jakarta.persistence.*;
 
 import java.util.List;
