+++ src/main/java/br/com/alura/AluraFake/task/model/TaskType.java
@@ -1,4 +1,4 @@
-package br.com.alura.AluraFake.task;
+package br.com.alura.AluraFake.task.model;
 
 public enum TaskType {
     OPEN_TEXT,
