+++ src/main/java/br/com/alura/AluraFake/user/UserController.java
@@ -1,5 +1,8 @@
 package br.com.alura.AluraFake.user;
 
+import br.com.alura.AluraFake.user.dto.NewUserDTO;
+import br.com.alura.AluraFake.user.dto.UserListItemDTO;
+import br.com.alura.AluraFake.user.model.User;
 import br.com.alura.AluraFake.util.ErrorItemDTO;
 import jakarta.validation.Valid;
 import org.springframework.http.*;
