+++ src/main/java/br/com/alura/AluraFake/user/UserRepository.java
@@ -1,6 +1,7 @@
 package br.com.alura.AluraFake.user;
 
 
+import br.com.alura.AluraFake.user.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
