+++ src/main/java/br/com/alura/AluraFake/user/dto/UserListItemDTO.java
@@ -1,4 +1,7 @@
-package br.com.alura.AluraFake.user;
+package br.com.alura.AluraFake.user.dto;
+
+import br.com.alura.AluraFake.user.model.Role;
+import br.com.alura.AluraFake.user.model.User;
 
 import java.io.Serializable;
