+++ src/main/java/br/com/alura/AluraFake/user/model/Role.java
@@ -1,4 +1,4 @@
-package br.com.alura.AluraFake.user;
+package br.com.alura.AluraFake.user.model;
 
 public enum Role {
     STUDENT,
