+++ src/test/java/br/com/alura/AluraFake/course/CourseControllerTest.java
@@ -1,8 +1,10 @@
 package br.com.alura.AluraFake.course;
 
+import br.com.alura.AluraFake.course.dto.NewCourseDTO;
+import br.com.alura.AluraFake.course.model.Course;
 import br.com.alura.AluraFake.task.TaskRepository;
-import br.com.alura.AluraFake.user.Role;
-import br.com.alura.AluraFake.user.User;
+import br.com.alura.AluraFake.user.model.Role;
+import br.com.alura.AluraFake.user.model.User;
 import br.com.alura.AluraFake.user.UserRepository;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.Test;
